+++ tests/src/Helper/Voter/AllowAllVoter.php
@@ -4,12 +4,13 @@ declare(strict_types=1);
 
 namespace WaffleTests\Commons\Security\Helper\Voter;
 
+use Waffle\Commons\Contracts\Auth\SecurityContextInterface;
 use Waffle\Commons\Contracts\Security\VoterInterface;
 
 final class AllowAllVoter implements VoterInterface
 {
     #[\Override]
-    public function decide(): bool
+    public function decide(SecurityContextInterface $ctx, mixed $subject = null): bool
     {
         return true;
     }